+++ src/tests/unit_tests/agent_tests/test_extract_seen_urls.py
@@ -4,12 +4,14 @@ from aidial_sdk.chat_completion import Message, Role
 from aidial_sdk.chat_completion.request import FunctionCall, ToolCall
 
 from quickapp.config.context import FileContextConfig, UserDefinedContextConfig
-from quickapp.internal_tooling.attachment_notification_tooling._tool_configs import (
-    AVAILABLE_CONTEXT_TOOL_NAME,
+from quickapp.internal_tooling.attachment_notification_tooling._context_entries import (
     ContextEntry,
     extract_seen_entries_from_messages,
     has_context_tool_history,
     should_activate_context_tool,
+)
+from quickapp.internal_tooling.attachment_notification_tooling._tool_configs import (
+    AVAILABLE_CONTEXT_TOOL_NAME,
 )
 
 TOOL_NAME = AVAILABLE_CONTEXT_TOOL_NAME
